requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
asyncer==0.0.2
bidict==0.23.1
cachetools==5.3.3
certifi==2024.7.4
chainlit==1.1.306
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
Deprecated==1.2.14
docstring_parser==0.16
fastapi==0.110.3
filetype==1.2.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.144.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.49.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
Lazify==0.4.0
literalai==0.0.607
marshmallow==3.21.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
packaging==23.2
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.1
pydantic_core==2.20.1
PyJWT==2.8.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
requests==2.32.3
rsa==4.9
shapely==2.0.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
syncer==2.0.3
tomli==2.0.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.24.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.25.0
vertexai==1.49.0
watchfiles==0.20.0
wrapt==1.16.0
wsproto==1.2.0
zipp==3.19.2
